main: Move SDL_RunApp bits from src/core to src/main.
Fixes #10170.
