+++ src/main/n3ds/SDL_sysmain_runapp.c
