+++ src/main/ngage/SDL_sysmain_runapp.cpp
