+++ src/main/winrt/SDL_sysmain_runapp.cpp
@@ -20,24 +20,21 @@
 */
 #include "SDL_internal.h"
 
+#include "../../core/winrt/SDL_winrtapp_direct3d.h"
+#include "../../core/winrt/SDL_winrtapp_xaml.h"
+
 #include <wrl.h>
 
-int (*WINRT_SDLAppEntryPoint)(int, char **) = NULL;
-
 extern "C"
-SDL_WinRT_DeviceFamily SDL_WinRTGetDeviceFamily()
+int SDL_RunApp(int, char**, SDL_main_func mainFunction, void * xamlBackgroundPanel)
 {
-#if NTDDI_VERSION >= NTDDI_WIN10 /* !!! FIXME: I have no idea if this is the right test. This is a UWP API, I think. Older windows should...just return "mobile"? I don't know. --ryan. */
-    Platform::String ^ deviceFamily = Windows::System::Profile::AnalyticsInfo::VersionInfo->DeviceFamily;
-
-    if (deviceFamily->Equals("Windows.Desktop")) {
-        return SDL_WINRT_DEVICEFAMILY_DESKTOP;
-    } else if (deviceFamily->Equals("Windows.Mobile")) {
-        return SDL_WINRT_DEVICEFAMILY_MOBILE;
-    } else if (deviceFamily->Equals("Windows.Xbox")) {
-        return SDL_WINRT_DEVICEFAMILY_XBOX;
+    if (xamlBackgroundPanel) {
+        return SDL_WinRTInitXAMLApp(mainFunction, xamlBackgroundPanel);
+    } else {
+        if (FAILED(Windows::Foundation::Initialize(RO_INIT_MULTITHREADED))) {
+            return 1;
+        }
+        return SDL_WinRTInitNonXAMLApp(mainFunction);
     }
-#endif
-
-    return SDL_WINRT_DEVICEFAMILY_UNKNOWN;
 }
+
